requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.3
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
numpy-utils==0.1.6
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pip==23.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
setuptools==68.2.0
six==1.17.0
wheel==0.41.2
zipp==3.21.0